Progress on traps.
More work on supporting traps, including work on x86
segmentation, supporting the IDT, TSS for stack tables,
and so on.

+++ theon/src/main.rs
@@ -25,7 +25,7 @@ mod x86_64;
 #[cfg_attr(not(test), no_mangle)]
 pub extern "C" fn main(mbinfo_phys: u64) -> ! {
     x86_64::init::start(mbinfo_phys);
-    panic!("main: trapstubs = {:#x}", arch::trap::stubs as u64);
+    panic!("main: trapstubs = {:#x}", arch::trap::stubs as usize);
 }
 
 #[no_mangle]
